PYTHON/New folder (2)/requirements.txt
google-generativeai
python-dotenv
streamlit
pillow